+++ BookStore.BuisneLogic/Models/OrderItems/OrderItemModelItem.cs
@@ -5,9 +5,9 @@ namespace BookStore.BusinessLogic.Models.OrderItems
 {
     public class OrderItemModelItem : BaseModel
     {
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public Currencys Currency { get; set; }
-        public int Count { get; set; }
+        public long Count { get; set; }
         public long OrderId { get; set; }
         public long PrintingEditionId { get; set; }
     }
