+++ BookStore.BuisneLogic/Models/PrintingEditions/PrintingEditionModelItem.cs
@@ -9,7 +9,7 @@ namespace BookStore.BusinessLogic.Models.PrintingEditions
         public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public int Price { get; set; }
+        public long Price { get; set; }
         public TypePrintingEditionEnum.Type Type { get; set; }
         public AuthorModel Authors { get; set; }
     }
