+++ BookStore.DataAccess/Entities/OrderItem.cs
@@ -6,14 +6,14 @@ namespace BookStore.DataAccess.Entities
 {
     public class OrderItem : BaseEntity
     {
-        public int Amount { get; set; }
+        public long Amount { get; set; }
         public Currencys Currency { get; set; }
-        public int Count { get; set; }
-        public int PrintingEditionId { get; set; }
+        public long Count { get; set; }
+        public long PrintingEditionId { get; set; }
         public PrintingEdition PrintingEdition { get; set; }
 
         [ForeignKey("OrderId")]
-        public int OrderId { get; set; }
+        public long OrderId { get; set; }
         public Order Order { get; set; }
     }
 }
