+++ BookStore.DataAccess/Entities/Payment.cs
@@ -4,6 +4,6 @@ namespace BookStore.DataAccess.Entities
 {
     public class Payment : BaseEntity
     {
-        public int TransactionId { get; set; }
+        public long TransactionId { get; set; }
     }
 }
